[FINAL] Tested
Implementation:
1) Download the ImageProcessor.hpp file and put it in the same folder as the robot.cpp file
2) Include the file by writing: #include "ImageProcessor.hpp"
Note: The include statement for the robot.hpp library in the robot.cpp file can be removed since it is already included in this class
3) Initialize an instance of the ImageProcessor class by writing "ImageProcessor imageProcessor;" inside the while loop
4) You should be able to access the values stored in the Result struct by writing "result.onLine" or "result.error" once the constructor has been called (step 3)
Note: result.onLine a boolean which is set to true if the robot is on the line and false if not. result.error is a double which contains the error margin (negative meaning the line is on the right of the cameraView's center and positive meaning the line is on the left of the cameraView's center). If result.onLine is set to false (the robot can not detect a line), result.error is set to 0.